move manager to integrations
